+++ 0x02-functions_nested_loops/103-fibonacci.c
@@ -24,11 +24,11 @@ void print_delim(void)
 int main(void)
 {
 	int fibo_count = 2;
-	long long int sum = 1;
-	long long int p1 = 1;
-	long long int p2 = 2;
+	unsigned long int sum = 1;
+	unsigned long int p1 = 1;
+	unsigned long int p2 = 2;
 
-	printf("%llu", p2);
+	printf("%lu", p2);
 	print_delim();
 
 	while (fibo_count < 50)
@@ -43,7 +43,7 @@ int main(void)
 		if (sum % 2 != 0)
 			continue;
 
-		printf("%lld", sum);
+		printf("%lu", sum);
 
 		if (fibo_count != 50)
 			print_delim();
